(refactor) Move player selection list to own component.
