+++ src/Main/PlayerSelecter.js
@@ -3,13 +3,10 @@ import PropTypes from 'prop-types';
 import { Link } from 'react-router';
 import ReactTooltip from 'react-tooltip';
 
-import AVAILABLE_CONFIGS from 'Parser/AVAILABLE_CONFIGS';
-import UnsupportedSpec from 'Parser/UnsupportedSpec/CONFIG';
-
-import makeAnalyzerUrl from './makeAnalyzerUrl';
 import DiscordButton from './DiscordButton';
 import PatreonButton from './PatreonButton';
 import GithubButton from './GithubButton';
+import PlayerSelectionList from './PlayerSelectionList';
 
 class PlayerSelecter extends Component {
   static propTypes = {
@@ -51,50 +48,7 @@ class PlayerSelecter extends Component {
             <h2>Select the player you wish to analyze</h2>
           </div>
           <div className="panel-body" style={{ padding: 0 }}>
-            <ul className="list selection players">
-              {combatants.length === 0 && (
-                <li className="text-danger" style={{ padding: '15px 22px' }}>
-                  Could not find any players in this report. Make sure the log is recorded with Advanced Combat Logging enabled. You can enable this in-game in the network settings.
-                </li>
-              )}
-              {
-                report.friendlies
-                  .sort((a, b) => {
-                    if (a.name > b.name) {
-                      return 1;
-                    } else if (a.name < b.name) {
-                      return -1;
-                    } else {
-                      return 0;
-                    }
-                  })
-                  .map(friendly => {
-                    const combatant = combatants.find(combatant => combatant.sourceID === friendly.id);
-                    if (!combatant) {
-                      return null;
-                    }
-                    let config = AVAILABLE_CONFIGS.find(config => config.spec.id === combatant.specID);
-                    if (!config) {
-                      if (process.env.NODE_ENV === 'development') {
-                        config = UnsupportedSpec;
-                      } else {
-                        return null;
-                      }
-                    }
-                    const spec = config.spec;
-                    const specClassName = spec.className.replace(' ', '');
-
-                    return (
-                      <li key={friendly.id} className="item selectable">
-                        <Link to={makeAnalyzerUrl(report, fightId, friendly.name)}>
-                          <img src={`/specs/${specClassName}-${spec.specName.replace(' ', '')}.jpg`} alt="Spec logo" />{' '}
-                          {friendly.name} ({spec.specName} {spec.className === 'Unsupported' ? `${spec.className} - ${friendly.type}` : spec.className})
-                        </Link>
-                      </li>
-                    );
-                  })
-              }
-            </ul>
+            <PlayerSelectionList report={report} fightId={fightId} combatants={combatants}/>
           </div>
         </div>
